+++ examples/User/JsonApi/Document/UsersDocument.php
@@ -45,6 +45,7 @@ class UsersDocument extends AbstractCollectionDocument
      */
     public function getLinks(): ?DocumentLinks
     {
-        return DocumentLinks::createWithoutBaseUri()->setPagination("/users", $this->object);
+        return DocumentLinks::createWithoutBaseUri()
+            ->setPagination("/users", $this->object, $this->request->getUri()->getQuery());
     }
 }