+++ src/JsonApi/Schema/Pagination/PaginationLinkProviderInterface.php
@@ -7,13 +7,13 @@ use WoohooLabs\Yin\JsonApi\Schema\Link\Link;
 
 interface PaginationLinkProviderInterface
 {
-    public function getSelfLink(string $uri): ?Link;
+    public function getSelfLink(string $uri, string $queryString): ?Link;
 
-    public function getFirstLink(string $uri): ?Link;
+    public function getFirstLink(string $uri, string $queryString): ?Link;
 
-    public function getLastLink(string $uri): ?Link;
+    public function getLastLink(string $uri, string $queryString): ?Link;
 
-    public function getPrevLink(string $uri): ?Link;
+    public function getPrevLink(string $uri, string $queryString): ?Link;
 
-    public function getNextLink(string $uri): ?Link;
+    public function getNextLink(string $uri, string $queryString): ?Link;
 }