+++ tests/JsonApi/Double/StubPaginationLinkProvider.php
@@ -8,27 +8,27 @@ use WoohooLabs\Yin\JsonApi\Schema\Pagination\PaginationLinkProviderInterface;
 
 class StubPaginationLinkProvider implements PaginationLinkProviderInterface
 {
-    public function getSelfLink(string $uri): ?Link
+    public function getSelfLink(string $uri, string $queryString): ?Link
     {
         return new Link($uri . "self");
     }
 
-    public function getFirstLink(string $uri): ?Link
+    public function getFirstLink(string $uri, string $queryString): ?Link
     {
         return new Link($uri . "first");
     }
 
-    public function getLastLink(string $uri): ?Link
+    public function getLastLink(string $uri, string $queryString): ?Link
     {
         return new Link($uri . "last");
     }
 
-    public function getPrevLink(string $uri): ?Link
+    public function getPrevLink(string $uri, string $queryString): ?Link
     {
         return new Link($uri . "prev");
     }
 
-    public function getNextLink(string $uri): ?Link
+    public function getNextLink(string $uri, string $queryString): ?Link
     {
         return new Link($uri . "next");
     }